Sometimes there is no history

+++ ci/scripts/tile-build.sh
@@ -10,7 +10,9 @@ BASE_DIR="$( cd "${REPO_DIR}/.." && pwd )"
 
 cd ${BASE_DIR}
 
-cp ${HISTORY_DIR}/tile-history.yml ${SOURCE_DIR}
+if [ -f "${HISTORY_DIR}/tile-history.yml" ]; then
+	cp ${HISTORY_DIR}/tile-history.yml ${SOURCE_DIR}
+fi
 (cd ${SOURCE_DIR}; ${REPO_DIR}/tile build)
 cp $${SOURCE_DIR}/product/*.pivotal ${TARGET_DIR}
 cp ${SOURCE_DIR}/tile-history.yml ${HISTORY_DIR}
